CHORE - vf-forms linting (#504)
* Minor style/spacing differences

* Add the generic JS template

* Make a JS module

* Namespace all the things, use data attributes
