+++ components/vf-form/vf-form__helper/vf-form__helper.scss
@@ -5,7 +5,7 @@
 
   color: set-color(vf-color--grey);
 
-  .floatLabel & {
+  .vf-form__floatlabel & {
     left: 24px;
     position: relative;
     top: 12px;
