+++ components/vf-form/vf-form__input/vf-form__input.njk
@@ -1,10 +1,10 @@
 <form action="" class="vf-form">
-  <div class="vf-form__item | floatLabel">
+  <div class="vf-form__item" data-vf-js-form-floatlabel>
     <label class="vf-form__label | vf-sr-only" for="text">Text</label>
     <input type="text" id="text" class="vf-form__input" placeholder="Text">
   </div>
 
-  <div class="vf-form__item | floatLabel">
+  <div class="vf-form__item" data-vf-js-form-floatlabel>
     <label class="vf-form__label | vf-sr-only" for="password">Password</label>
     <input type="password" id="password" class="vf-form__input" placeholder="Password">
   </div>
