+++ components/vf-form/vf-form__input/vf-form__input.scss
@@ -67,33 +67,33 @@
   color: set-color(vf-color--grey--dark);
 }
 
-.vf-form__input select.vf-form__select {
+.vf-form__input .vf-form__select {
   appearance: none;
   color: set-color(vf-color--grey--dark);
   cursor: pointer;
   position: relative;
 }
 
-.vf-form__input select.vf-form__select::-ms-expand {
+.vf-form__input .vf-form__select::-ms-expand {
   display: none;
 }
 
-.vf-form__input .vf-form__is-active input.vf-form__input,
-.vf-form__input .vf-form__is-active select.vf-form__select,
-.vf-form__input .vf-form__is-active textarea.vf-form__textarea {
+.vf-form__input .vf-form__is-active .vf-form__input,
+.vf-form__input .vf-form__is-active .vf-form__select,
+.vf-form__input .vf-form__is-active .vf-form__textarea {
   background-color: set-ui-color(vf-ui-color--white);
   border-color: set-ui-color(vf-ui-grey);
   color: set-color(vf-grey-dark);
 }
 
-.vf-form__input .vf-form__has-focus input.vf-form__input,
-.vf-form__input .vf-form__has-focus select.vf-form__select,
-.vf-form__input .vf-form__has-focus textarea.vf-form__textarea {
+.vf-form__input .vf-form__has-focus .vf-form__input,
+.vf-form__input .vf-form__has-focus .vf-form__select,
+.vf-form__input .vf-form__has-focus .vf-form__textarea {
   background-color: set-ui-color(vf-ui-color--white);
   border-color: set-ui-color(vf-ui-color--black);
 }
 
-.vf-form__wrap .vf-form__input label.vf-form__label {
+.vf-form__wrap .vf-form__input .vf-form__label {
   border-radius-top-left: 3px;
   color: set-color(vf-color--grey--dark);
   display: block;
@@ -106,12 +106,12 @@
   z-index: 1;
 }
 
-.vf-form__input .vf-form__is-active label.vf-form__label {
+.vf-form__input .vf-form__is-active .vf-form__label {
   opacity: 1;
   visibility: visible;
 }
 
-.vf-form__input .vf-form__has-focus label.vf-form__label {
+.vf-form__input .vf-form__has-focus .vf-form__label {
   color: set-ui-color(vf-ui-color--black);
 }
 
@@ -139,37 +139,11 @@
 * modify it as you wish!
 */
 
-.vf-form__item.floatLabel {
+.vf-form__item.vf-form__floatlabel {
   position: relative;
 }
-.floatLabel {
+.vf-form__floatlabel {
   .vf-form__label {
     position: absolute;
   }
-}
-
-/* Start CSS3 Animations on supported browser */
-.vf-form__item.floatLabel .placeHolder {
-  left: 0;
-  opacity: 0;
-  position: absolute;
-  top: 0;
-  transition-delay: 100ms;
-  transition: opacity 120ms, visibility 120ms;
-  visibility: hidden;
-  white-space: nowrap;
-  width: auto;
-}
-
-
-.vf-form__item.floatLabel .placeHolder.active {
-  display: block;
-  left: .5em;
-  opacity: 1;
-  position: absolute;
-  top: 0;
-  transition: opacity 120ms, visibility 120ms;
-  transition-delay: 100ms;
-  visibility: visible;
-  width: auto;
 }
