+++ components/vf-form/vf-form__item/vf-form__item.njk
@@ -1,4 +1,4 @@
-<div class="vf-form__item | floatLabel">
+<div class="vf-form__item" data-vf-js-form-floatlabel>
   <label for="text" class="vf-form__label | vf-sr-only">Form Label</label>
   <input type="text" id="text" class="vf-form__input" placeholder="for item example">
   <p class="vf-form__helper">Form helper text</p>
