+++ components/vf-form/vf-form__textarea/vf-form__textarea.njk
@@ -1,4 +1,4 @@
-<div class="vf-form__item | floatLabel">
+<div class="vf-form__item" data-vf-js-form-floatlabel>
   <label class="vf-form__label | vf-sr-only" for="vf-form__textarea">Write Some More details</label>
   <textarea class="vf-form__textarea" id="vf-form__textarea" name="vf-form__textarea" rows="5" placeholder="Write Some More details"></textarea>
 </div>
