+++ components/vf-masthead/vf-masthead.njk
@@ -9,7 +9,7 @@
         <span class="vf-masthead__group">Structural Biology</span>
       </h2>
     </div>
-    
+
     <form action="" class="vf-masthead__form--search">
       <div class="vf-masthead__form__item">
         <label for="" class="vf-masthead__form__label">Search This Project</label>
